% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Printer Function}
\usage{
printer(r, x)
}
\arguments{
\item{r}{What in the second column}

\item{x}{What in the first column}
}
\value{
A tibble
}
\description{
Printer Function
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
